opt: move logging module into internal package
Avoid exposing some unwanted variables and methods to users.
