+++ gnet.go
@@ -27,7 +27,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/panjf2000/gnet/logging"
+	"github.com/panjf2000/gnet/internal/logging"
 )
 
 // Action is an action that occurs after the completion of an event.
