+++ internal/netpoll/epoll.go
@@ -28,7 +28,7 @@ import (
 	"unsafe"
 
 	"github.com/panjf2000/gnet/internal"
-	"github.com/panjf2000/gnet/logging"
+	"github.com/panjf2000/gnet/internal/logging"
 	"golang.org/x/sys/unix"
 )
